update_dns/pyproject.toml
[project]
name = "update_dns"
version = "0.1.0"
description = "Dynamic DNS updater for VPN using CloudFlare API"
authors = [
    {name = "Brian Kaewell", email = "briankaewell@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
packages = [{ include = "update_dns", from = "src" }]
requires-python = ">=3.10, <4.0"
dependencies = [
    "gspread (>=6.2.1, <7.0.0)",
    "oauth2client (>=4.1.3, <5.0.0)",
    "python-dotenv (>=1.1.1, <2.0.0)",
    "requests (>=2.32.5, <3.0.0)",
]

#[tool.poetry.scripts]
#Console script entry point
#update-vpn-ddns = "update_vpn_ddns.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0, <3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
responses = "^0.25.8"
pytest-cov = "^6.3.0"
black = "^25.1.0"
